VBA-Challenge Stock Market Analysis
Create a script that loops through all the stocks and outputs

The ticker symbol in the "Ticker" Column

Yearly change from the opening price at the beginning of a given year to the closing price at the end of that year with conditional formatting to express positive change in green and negative change in red.

The percentage change from the opening price at the beginning of a given year to the closing price at the end of that year.

The total stock volume of the stock.

The return stock with the "Greatest % increase", "Greatest % decrease", and "Greatest total volume"